move query plan struct definitions to another module
